llm-server/requirements.txt
aiohttp==3.9.3
beanie==1.25.0
boto3==1.34.64
botocore==1.34.64
fastapi==0.110.1
fastapi_admin==1.0.4
httpx==0.27.0
motor==3.3.2
passlib==1.7.4
pydantic==2.6.4
python-dotenv==1.0.1
python_jose==3.3.0
redis==5.0.3
Requests==2.31.0
starlette==0.37.2
uvicorn==0.29.0
